Updated code for cart.

+++ app/src/main/java/com/appyhome/appyproduct/mvvm/data/model/api/product/AddToCartRequest.java
@@ -7,17 +7,17 @@ import com.google.gson.annotations.SerializedName;
 public class AddToCartRequest {
     @Expose
     @SerializedName("product_id")
-    public int product_id;
+    public long product_id;
 
     @Expose
     @SerializedName("variant_id")
-    public int variant_id;
+    public long variant_id;
 
     @Expose
     @SerializedName("quantity")
     public int quantity;
 
-    public AddToCartRequest(int idP, int idV, int amount) {
+    public AddToCartRequest(long idP, long idV, int amount) {
         quantity = amount;
         product_id = idP;
         variant_id = idV;
