+++ app/src/main/java/com/appyhome/appyproduct/mvvm/data/model/api/product/DeleteCartRequest.java
@@ -7,13 +7,13 @@ import com.google.gson.annotations.SerializedName;
 public class DeleteCartRequest {
     @Expose
     @SerializedName("product_id")
-    public int product_id;
+    public long product_id;
 
     @Expose
     @SerializedName("variant_id")
-    public int variant_id;
+    public long variant_id;
 
-    public DeleteCartRequest(int idP, int idV) {
+    public DeleteCartRequest(long idP, long idV) {
         product_id = idP;
         variant_id = idV;
     }
