+++ app/src/main/java/com/appyhome/appyproduct/mvvm/data/model/api/product/EditCartRequest.java
@@ -7,17 +7,17 @@ import com.google.gson.annotations.SerializedName;
 public class EditCartRequest {
     @Expose
     @SerializedName("product_id")
-    public int product_id;
+    public long product_id;
 
     @Expose
     @SerializedName("variant_id")
-    public int variant_id;
+    public long variant_id;
 
     @Expose
     @SerializedName("new_quantity")
     public int new_quantity;
 
-    public EditCartRequest(int idP, int idV, int amount) {
+    public EditCartRequest(long idP, long idV, int amount) {
         new_quantity = amount;
         product_id = idP;
         variant_id = idV;
